Vendor icons on the pull request that declares them
Following the screenshot workflow, a contributor should not have to fetch
and commit an icon by hand just because they declared where it lives. Add
a pull_request_target job that syncs the icons for the Blueprints the
branch touches and commits the result to the branch, guarded to same-repo
pull requests exactly as the screenshot job is.

Only the touched Blueprints are synced, so an unrelated upstream change is
never smuggled into somebody else's pull request. The scheduled job keeps
syncing everything, since it exists to catch icons that changed upstream
after their pull request merged.

Move getChangedFiles()/getTouchedBlueprintDirectories() into
scripts/lib/changed-files.js, shared with validate-pr-blueprints.js rather
than duplicated.
