+++ .github/workflows/sync-app-icons.yml
@@ -4,14 +4,76 @@ on:
     workflow_dispatch:
     schedule:
         - cron: "0 5 * * 1"
+    pull_request_target:
+        types: [opened, synchronize, reopened]
+        paths:
+            - "blueprints/**/app-meta.json"
+            - ".github/workflows/sync-app-icons.yml"
+            - "scripts/sync-app-icons.js"
 
 permissions:
     contents: write
     pull-requests: write
 
 jobs:
-    sync:
+    # Vendor the icon for the Blueprints a pull request touches, so declaring
+    # an iconSource is all a contributor has to do.
+    sync_pull_request:
         runs-on: ubuntu-latest
+        # Only run for PRs from the same repository (not forks)
+        if: |
+            github.event_name == 'pull_request_target' &&
+            github.event.pull_request.head.repo.full_name == github.repository
+        env:
+            GITHUB_BASE_REF: ${{ github.event.pull_request.base.ref }}
+        steps:
+            - name: Checkout PR branch
+              uses: actions/checkout@v4
+              with:
+                  ref: ${{ github.event.pull_request.head.ref }}
+                  token: ${{ secrets.GITHUB_TOKEN }}
+                  fetch-depth: 0
+
+            - name: Fetch base branch
+              run: |
+                  git fetch origin "${GITHUB_BASE_REF}:refs/remotes/origin/${GITHUB_BASE_REF}"
+
+            - name: Collect changed files
+              run: |
+                  set -euo pipefail
+                  merge_base="$(git merge-base "origin/${GITHUB_BASE_REF}" HEAD)"
+                  git diff --name-only "${merge_base}" HEAD > "${RUNNER_TEMP}/changed-files.txt"
+                  echo "CHANGED_FILES_PATH=${RUNNER_TEMP}/changed-files.txt" >> "${GITHUB_ENV}"
+
+            - uses: actions/setup-node@v4
+              with:
+                  node-version: 20
+
+            - name: Sync icons for the touched Blueprints
+              run: node scripts/sync-app-icons.js --changed-only
+
+            - name: Check for changes
+              id: check_changes
+              run: |
+                  if [ -n "$(git status --porcelain)" ]; then
+                    echo "has_changes=true" >> $GITHUB_OUTPUT
+                  else
+                    echo "has_changes=false" >> $GITHUB_OUTPUT
+                  fi
+
+            - name: Commit and push to PR
+              if: steps.check_changes.outputs.has_changes == 'true'
+              run: |
+                  git config --global user.name "github-actions[bot]"
+                  git config --global user.email "github-actions[bot]@users.noreply.github.com"
+                  git add blueprints
+                  git commit -m "chore: sync app icons from upstream"
+                  git push
+
+    # Catch the icons that changed upstream after their pull request merged.
+    sync_trunk:
+        runs-on: ubuntu-latest
+        if: github.event_name != 'pull_request_target'
         steps:
             - uses: actions/checkout@v4
