+++ scripts/validate-pr-blueprints.js
@@ -1,6 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
 
+import { getTouchedBlueprintDirectories } from './lib/changed-files.js';
 import {
 	ajvPath,
 	createBlueprintValidator,
@@ -20,34 +21,6 @@ function getPullRequestSource() {
 	}
 
 	return { repository, ref };
-}
-
-function getChangedFiles() {
-	if (process.env.CHANGED_FILES) {
-		return process.env.CHANGED_FILES.split(/\r?\n/).filter(Boolean);
-	}
-
-	if (process.env.CHANGED_FILES_PATH) {
-		return fs
-			.readFileSync(process.env.CHANGED_FILES_PATH, 'utf8')
-			.split(/\r?\n/)
-			.filter(Boolean);
-	}
-
-	throw new Error('CHANGED_FILES_PATH or CHANGED_FILES must be provided.');
-}
-
-function getTouchedBlueprintDirectories() {
-	const blueprintDirs = new Set();
-
-	for (const changedFile of getChangedFiles()) {
-		const match = changedFile.match(/^(blueprints\/[^/]+)(?:\/|$)/);
-		if (match) {
-			blueprintDirs.add(match[1]);
-		}
-	}
-
-	return [...blueprintDirs].sort();
 }
 
 function isLocalFile(filePath) {
